Move extension-types.ts to the store
