+++ src/frontend/packages/core/src/features/endpoints/connect.service.ts
@@ -19,12 +19,12 @@ import { STRATOS_ENDPOINT_TYPE } from '../../../../store/src/base-entity-schemas
 import { EndpointsEffect } from '../../../../store/src/effects/endpoint.effects';
 import { SystemEffects } from '../../../../store/src/effects/system.effects';
 import { entityCatalog } from '../../../../store/src/entity-catalog/entity-catalog';
+import { EndpointType } from '../../../../store/src/extension-types';
 import { endpointSchemaKey } from '../../../../store/src/helpers/entity-factory';
 import { ActionState } from '../../../../store/src/reducers/api-request-reducer/types';
 import { selectEntity, selectRequestInfo, selectUpdateInfo } from '../../../../store/src/selectors/api.selectors';
 import { EndpointModel } from '../../../../store/src/types/endpoint.types';
 import { EndpointsService } from '../../core/endpoints.service';
-import { EndpointType } from '../../core/extension/extension-types';
 import { safeUnsubscribe } from '../../core/utils.service';
 
 export interface ConnectEndpointConfig {
