+++ src/frontend/packages/core/src/features/endpoints/endpoint-auth.ts
@@ -1,9 +1,10 @@
-import { EndpointAuthTypeNames } from './endpoint-helpers';
 import { Validators } from '@angular/forms';
-import { EndpointType, EndpointAuthTypeConfig } from '../../core/extension/extension-types';
+
+import { EndpointAuthTypeConfig, EndpointType } from '../../../../store/src/extension-types';
 import { CredentialsAuthFormComponent } from './connect-endpoint-dialog/auth-forms/credentials-auth-form.component';
-import { SSOAuthFormComponent } from './connect-endpoint-dialog/auth-forms/sso-auth-form.component';
 import { NoneAuthFormComponent } from './connect-endpoint-dialog/auth-forms/none-auth-form.component';
+import { SSOAuthFormComponent } from './connect-endpoint-dialog/auth-forms/sso-auth-form.component';
+import { EndpointAuthTypeNames } from './endpoint-helpers';
 
 export abstract class BaseEndpointAuth {
   static readonly UsernamePassword = {
