+++ src/frontend/packages/store/src/entity-catalog/entity-catalog.types.ts
@@ -2,7 +2,6 @@ import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
 import { EndpointHealthCheck } from '../../../core/endpoints-health-checks';
-import { EndpointAuthTypeConfig } from '../../../core/src/core/extension/extension-types';
 import { FavoritesConfigMapper } from '../../../core/src/shared/components/favorites-meta-card/favorite-config-mapper';
 import { StratosStatus } from '../../../core/src/shared/shared.types';
 import { GeneralEntityAppState } from '../app-state';
@@ -21,6 +20,7 @@ import {
 import {
   PaginationPageIteratorConfig,
 } from '../entity-request-pipeline/pagination-request-base-handlers/pagination-iterator.pipe';
+import { EndpointAuthTypeConfig } from '../extension-types';
 import { EntitySchema } from '../helpers/entity-schema';
 import { UserFavorite } from '../types/user-favorites.types';
