+++ src/frontend/packages/store/src/types/endpoint.types.ts
@@ -1,6 +1,6 @@
-import { EndpointType } from '../../../core/src/core/extension/extension-types';
 import { StratosScopeStrings } from '../../../core/src/core/permissions/stratos-user-permissions.checker';
 import { MetricsAPITargets, MetricsStratosInfo } from '../actions/metrics-api.actions';
+import { EndpointType } from '../extension-types';
 import { endpointSchemaKey } from '../helpers/entity-factory';
 import { RequestSectionKeys, TRequestTypeKeys } from '../reducers/api-request-reducer/types';
